requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
pydantic>=1.8.0,<2.0.0
openai>=0.27.0,<0.28.0
qdrant-client>=1.1.0,<1.2.0
python-dotenv>=0.19.0,<0.20.0
